luci-app-acme: DNS API: list of all providers
To simplify configuration we need a list of all supported providers.
Each provider has own set of params. We should show them when a user selected a provider.
To help a user to find instruction for the provider we will generate a link to Wiki.

A user still can change the Credentials option manually.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>