luci-app-lldpd: Create from Tano Systems
This is a copy of https://github.com/tano-systems/luci-app-tn-lldpd, which is licensed under the MIT License.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Rename luci-app-tn-lldpd to luci-app-lldpd

The original version of Tanosystem has a naming scheme which does not correspond to the standard naming scheme in OpenWrt LuCi. Therefore the renaming.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Fix bug not getting the current lldpd status

Specifying the arguments in the wrong order can (and has in my tests) resulted in errors. So you can see in the man page that flags like -f come first and then the command.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Disable option to enable SNMP agent

SNMP agent support is not enabled by default in lldpd. This can (and has in my tests) cause LLDP to stop working. See comment in source code.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Add option to disable sending sensitive information

Added an option that sets the "-k" flag, which results in less sensitive information being sent. See man pages and description in source code.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Add license information to Makefile

All of Tano Systems source code for the app is licensed under the MIT License. This has now been indicated accordingly in the Makefile.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Remove old footer from Tano Systems

The app from Tano Systems appears to include a footer. Since most LuCi apps do not, I have removed it.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Bug fix caused that the management IP address could not be set.

The TanoWrt fork calls the option to set the management IP addresses "lldp_sys_mgmt_ip". However, in OpenWrt it is called "lldp_mgmt_ip".

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Improve style

Remove double space

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Adding the Lua dependency

The rpcd script used by the Luci app is written in Lua. Therefore, runtime dependencies for Lua must exist when using the Luci app.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Remove old files

Since the app is called luci-app-lldpd and not luci-app-tn-lldpd, the files for the TanoWrt app are no longer necessary.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: RPCd backend change from Lua to ucode

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Update filter matrix

The old filter matrix was based on the man pages, but the selected value in the range 1-20 was also assigned to the configuration and thus lldpd was configured incorrectly. lldpd expects a value in the range 0-19. The webapp matrix has now been changed so that the values 0-19 also appear there.
This leads to a unification of configuration and webapp.

Signed-off-by: Marek Küthe <m.k@mk16.de>

luci-app-lldpd: Add location parameter

Signed-off-by: Marek Küthe <m.k@mk16.de>