luci-app-acme: Reuse translations for removed state_dir option
In the commit acf40dc0d402c68f35c02982a119e4a64131ce04 was removed the state_dir.
But translations were kept and they still have a value.
We should say to a user where the certs are stored now, and we can reuse the old translations.
They need to be fixed a little bit and already fixed for Russian, Ukrainian and Bulgarian languages.

See #6273

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>