collections: add luci-lighttpd
Some firmwares like Turris and Gl.Inet use Lighttpd for Luci.

The mod_alias is often used so we need to set luci alias paths.

The lighttpd use own user but for Luci we need root user.
That's why the server.username is overridden to empty.
When removing the package it will be rolled back to use the default lighttpd user.
But the log file was created with root.
We should remove it manually to allow the lighttpd user to recreate it.

We also need to enable CGI for the cgi-bin folder.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>