luci-mod-status: add missing vht40 channels
As of now, channel_analysis detects

- regular 20 MHz channels (all kinds: it is the default)
- 40 MHz HT channels, as well as
- 80 and 80+80 and 160 MHz VHT channels.

It is missing 40 MHz VHT channels. What it does for those right now is
display them as 20 MHz channels (both in the plot and the table below),
but it uses the center frequency of the 40 MHz channel, which means
their (wrong) 20 MHz band in the plot is offset by 2.

This fixes #6419. Note that this only occurs for **other** stations.
The code for local_wifi is different and already complete.

Signed-off-by: Frank Löffler <frank.loeffler@uni-jena.de>