luci-mod-status: correct 40 MHz intolerant channel
Use channel width info from HT for reported channel VHT channel
widths <=40. This in particular takes care of 40 MHz intolerant
stations that nevertheless report 40 Mhz width, but do not define
a secondary channel (and thus, are effectively 20 MHz). Such a
station will have set VHT to be set to effectively "20 MHz or 40 MHz"
(there is no VHT state that would specify which), and since it
reports HT 40 MHz width (but without specifying a secondary channel),
it will also report VHT 40 MHz, but with the main (20 MHz) frequency
as the center_freq_1, which in this case is misleading. What we have
to do instead of relying on the center_freq_1 of VHT for
bandwidths<=40 is to look at HT, as only there we find the information
whether this is in fact width 20 or 40 Mhz - and there the correct
center frequency will have been setup as well already, so this can
be a no-op in the code.

Signed-off-by: Frank Löffler <frank.loeffler@uni-jena.de>