luci-mod-status: channel_analysis: extend comment
Extend code comment about why the code block for vht does not and
should not deal with bandwidth <= 40:

The problem are stations that announce to be 40 MHz-capable, but only
use 20 MHz. There, ht_operation.channel_width == 2040 and
vht_operation.channel_width==40. The latter might look odd, but the VHT
mode only contains a bit field for "20 or 40 MHz" and cannot distinguish
between those, so the '40' there really mean "<=40". The HT-branch
before checks for this combination and calls it "20 MHz (40 MHz
Intolerant)". It does so by looking at
res.ht_operation.secondary_channel_offset. There is no such field for
VHT: one has to look at HT instead in that case, as also for VHT, a
bandwidth of 40 MHz (really: <=40 MHz) is announced, but only 20 MHz are
used (VHT cannot announce 20 MHz). This is what the patch wants to fix:
to prevent channel_analysis to use the 20 MHz-frequency in
res.vht_operation.center_freq_1 together with a (wrong) value of 40 MHz
for the bandwidth.

Whether 40 MHz are included or not in the VHT "if" only matters if there
would be a station where VHT is enabled with a bandwidth<=40, but HT is
not enabled at all. Can there be such a station (VHT is on, but HT is
not)? If such a station could exist, we would have a problem if that
station uses a bandwidth <=40 MHz: we would have no way of knowing the
actual bandwidth of the station: we would only know it to be <=40 MHz.
See
https://github.com/openwrt/iwinfo/blob/a34977c0760c93480491c8eb94da656b57d7f4cc/include/iwinfo.h#L311

Signed-off-by: Frank Löffler <frank.loeffler@uni-jena.de>