luci-mod-status: channel_analysis: correct VHT 20/40 MHz center channels (#6728)
* luci-mod-status: add missing vht40 channels

As of now, channel_analysis is missing 40 MHz VHT channels. 
What it does for those right now is
display them as 20 MHz channels (both in the plot and the table below),
but it uses the center frequency of the 40 MHz channel, which means
their (wrong) 20 MHz band in the plot is offset by 2.

This fixes #6419. Note that this only occurs for **other** stations.
The code for local_wifi is different and already complete.

Signed-off-by: Frank Löffler <frank.loeffler@uni-jena.de>