treewide: Add CODEOWNERS and fix missing PKG_MAINTAINER
The CODEOWNERS will request for a reviewer automatically by a path.

https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners

The luci-app-bcp38 had a maintainer from luci-app-acme and that looks like a copy-paste error.
So it's maintainer was changed to an author @danrl

Daniel F. Dickinson changed email address to <dfdpublic@wildtechgarden.ca>

luci-all-lxl has a maintainer Petar Koretic <petar.koretic@sartura.hr> but there is no corresponding GitHub account.
So Dirk Brenken was added as a second maintainer: he answered on an issue of the app.

When maintainer wasn't set the initial author was used, or most contributor or Jo-Philipp Wich as a default.

If anyone can't be a maintainer it should create an issue and the CODEOWNERS file will be updated with a new one.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>