luci-mod-network: put each IPv6 address on a separate line
Right now, when there's active DHCPv6 leases and more than one address per host
(e.g. because ULA is used together with GUA prefixes), the two IPv6 addresses
will be printed on one line (which may or may not get broken up depending on
the length of the addresses, which also looks inconsistent).

Putting each address on a separate line makes it much easier to read the
addresses (IMHO).

Signed-off-by: David Härdeman <david@hardeman.nu>