luci-app-tor: add config for tor
Currently the UCI for the Tor allows to set only a list of configs to include.
The only way to configure it for a user is to upload its own config and add it to list of included.
We can simplify this with the app.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>
(cherry picked from commit c170dfab9197f714c5e4483caad924631779122a)