luci-app-opkg: add warning about modifying packages
It is a well-known fact that opkg doesn't check ABI-compatibility between dependencies and upgrading packages in a stable release is therefore not recommended. In addition, installing and removing packages blindly may result in rendering the device inaccessible as well, so caution should be taken when doing this.

This commit adds a warning which should help notify especially new users about these behaviors with a link to the wiki which advises to not upgrade packages.

Signed-off-by: Daniel Nilsson <daniel.nilsson94@outlook.com>