luci-mod-network: Add clarifying note to wireless join dialog
Add a clarifying note about two different meanings of "network"on
the wireless join page. In the dialog title the word "network" refers
to the wireless SSID, while in the config item the same "network"
refers to the name to be used in OpenWrt network configuration.

Reference to forum discussion at
https://forum.openwrt.org/t/cant-join-a-scanned-network/189859

Signed-off-by: Hannu Nyman <hannu.nyman@iki.fi>