luci-base: fix license
During the update of luasrcdiet with commit b5d5e5bf13e572af7f1589a01bfd338691d5f063
the license was incorrectly changed from 'Apache-2.0' to 'MIT'.

As the name implies, this is the basic package that must be installed on the
target for the LuCI to work at all. So changing the license from 'Apache-2.0'
to 'MIT' should not have happend, because this information belongs to the
package 'luci-base' and to the 'luasrcdiet'.

Therefore, with commit b0d8fff3e40568af633032aacb0624da16f4266c the
'luasrcdiet' was moved to its own package. However, it was forgotten to adapt
the licence to 'Apache-2.0' again.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>