luci-base: don't render until luci-loaded is done
Because the setupDOM/initDOM methods do strange things related
to the old lua server rendered templates
(like remove all the elements currently hidden by dependencies...),
we want to be sure that these have finished before the view
itself renders.

This also ensures that any caching (e.g. from probeSystemFeatures)
is finished before the view render.

This feels like a hacky fix, but it's not clear to me what the best
one is.

Signed-off-by: James Haggerty <james.haggerty@morsemicro.com>