luci-mod-network: add support for bonding device in netifd
netifd support for bonding devices has been added in commit 5ba9744.
The other way to have bonding available in luci is via
luci-proto-bonding and its proto-bonding dependency.

luci-proto-bonding requires an IP address on a bond interface due to a
netifd limitation. This can cause some issues if you want to bridge an
Ethernet port and a bond.

This patch aims to have the device to be configurable not as an
interface protocol but as a device.

Signed-off-by: Nicolò Veronese <nicveronese@gmail.com>