luci-app-acme: Remove use_acme_server option since it's ignored by acme
In the 1ad72c6fb5d1c49147194de4f1bca1eb9d1f8f88 was added use_acme_server flag and acme_server URL.
The problem is that the acme uses the only acme_server option.
A user may think that once the flag is turned off it will be ignored, but it's not.

To avoid confusion, it's easier just to remove the flag.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>