luci-mod-system: uhttpd: add listen_http, listen_https, and enabled options
A user may want to change the uhttpd port or IP address e.g. set it to 192.168.1.1.
If the user switched to another web server e.g. nginx or lighttpd he may also want to stop the uhttpd.
The settings normally shouldn't be changed, but if a user starts to use a second web server then it will be enough to change only these options.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>