luci-app-attendedsysupgrade: use remote revision for version code
With 8860ca069c04e3dbd3f107f1526f2901f1085902 the check against the version_code
parameter was implemented but was incorrectly set to the current revision
instead of the latest available remote revision.

This resulted in not being able to upgrade due to the version_code parameter
never matched the newer remote revision on the sysupgrade server.

Signed-off-by: Daniel Nilsson <daniel.nilsson94@outlook.com>