luci-app-attendedsysupgrade: use versioned endpoints
The unversioned endpoint /api/overview has been removed from ASU which breaks
the request when searching for an available sysupgrade if the device is on a
stable release. In the case where the unversioned endpoint is added back to
prevent this regression in 23.05.3 and earlier (so upgrade is once again
possible) we should anyway rely on the versioned endpoints as those replace
the unversioned ones.

Signed-off-by: Daniel Nilsson <daniel.nilsson94@outlook.com>