luci-mod-network: split interface name and protocol validation
The previous implementation of binding the protocol validation to the
name validation caused problems when creating a new interface where an
already existing interface had the new one's protocol as its name, as the
protocol would be used when validating if an interface of that name already
existed.

Consider the following case:
1. Interface 'gre' with protocol PPPoE is created
2. Interface 'foo' with protocol GRE tunnel over IPv4 (which resolves to
   'gre') tries to be created

Creating interface 'foo' would error out during protocol validation as
'gre' would be passed to the name validation, which in turn would check if
an interface with name 'gre' already exists, which it does.

Instead of reusing the validation logic, simply trigger the name
validation manually as the protocol changes to properly pass the name of
the interface instead of the protocol. This also gives the benefit of
keeping all error states related to the name contained in the appropriate
UI element.

Fixes #7146

Tested-by: Tiago Gaspar <tiagogaspar8@gmail.com>
Signed-off-by: Daniel Nilsson <dannil+github@protonmail.com>