luci-app-clamav: Convert to JS
Added functionality for clamav-milter and freshclam which are also in
the package.

Assumptions: clamd.log will always be in /tmp, so read permissions are
granted for /tmp/*, in case the log file gets moved or renamed.

Don't expose the LogFile* parameters to avoid logfile related location
problems.

Added entries to use syslogging for each component, and added a Log tab.

Tested on 23.05.5

Signed-off-by: Paul Donald <newtwen+github@gmail.com>