luci-app-squid: convert to JavaScript
The conversion adds the missing parameters in the UI which is
configurable in the principal package.

Assumption has been made that the config file is stored in /etc/squid
since the principal package limits the sysconfdir to this directory. If
that assumption is changed in the future we need to adjust the ACL.

Signed-off-by: Daniel Nilsson <dannil+github@protonmail.com>