luci-mod-system: Allow selective file restore from backups
This is useful when you're importing backups intended for different
hardware, for example, where e.g. `/etc/config/[network|wireless]` are
not suitable for the target platform, but everything else useful.

Caveats: subsequent extraction and compression steps happen on the
target device, to then feed the selective_backup archive to sysupgrade.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>