luci-app-opkg: fix passing wrong option on opkg update/install
Fix passing wrong option on opkg update/install. While starting to
introduce support for APK in the opkg module, it was notice that
--force-removal-of-dependent-packages was always passed even with update
and install command.

This was probably a leftover/oversight of old one. To fix this, limit
this option only on remove and also update the acl.d to support single
call to update or install.

Fixes: 9b25031cb29b ("luci-app-opkg: rework backend operations")
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>