luci-app-openvpn: Handle missing openvpn config file.
With this change if the specified config file (from UCI) doesn't exist,
it won't error out. instead it treats it as an empty file.
Previously, when clicking edit on the custom_config option in the UI, it
will show an error because the file doesn't exist there by default. As a
result, a user can't easily add (paste) a config file content.

Signed-off-by: Milad Mohtashamirad <milad.mohtashamirad@morsemicro.com>