luci-base: prevent trace-back loading wireless on non-WiFi devices
Manifests in the console as:

RPCError: RPC call to uci/get failed with ubus code 4: Resource not found

The lazy-loading is harmless and has been that way for years, but the
presence of the trace-back clouds other problems. This should keep the
console cleaner; it conditionally attempts to load wireless when a more
fundamental check for it passes.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>