luci-app-attendedsysupgrade: add firmware upgrade check status include
The status include is named '11' so as to come immediately after
'10_system' (which displays the system properties). Its only permanent
addition to the status page is a checkbox for the user preference to
check for updates.

If the UCI luci core section contains the option:
  `check_for_newer_firmwares`
then upon load, the status page checks if the download server offers a
new(er) firmware compatible with the current device.

Signed-off-by: Paul Spooren <mail@aparcar.org>
Signed-off-by: Paul Donald <newtwen+github@gmail.com>