luci-base: revert timeout function argument for addNotification
The function signature changed and all function call to addNotification
that add a CSS class does not work anymore. A direct revert is not
possible.

If this feature is needed than a new function musst be added, that does
not break the function signature.

Fixes: 53e36e3293bee2ea2ce0fbc1250074c44cfe8335

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit 3f832f25120184a5794f593bafa762747d6cc693)