luci-app-attendedsysupgrade: maintain value of owut.rootfs_size in LuCI app
If someone has used owut to install an image with a modified rootfs
partition size, then using LuCI Attended Sysupgrade would change
it back, causing either a build failure or loss of configuration.

We fix this by simply adding the value from the owut config to the
build request (or 'null' if it doesn't exist).  The ASU server
then builds the image with the correct rootfs size, avoiding
these issues.

This value is not and should not be exposed in the LuCI app's
options, as it is quite advanced, very fragile and should only be
used by those who have researched the issues involved, typically
by reading the owut documentation and testing with owut first.

Signed-off-by: Eric Fahlgren <ericfahlgren@gmail.com>