luci-app-ddns: no implicit interface when ip_source=interface
The DDNS service attribute "interface" is for hotplug.d hook to bring up
and down DDNS updater(s) on ifup and ifdown events for an OpenWrt
"interface", not to be confused with attribute "ip_interface" which defines
the physical interface to get IP from.

With this change, the DDNS configuration service.xxx.interface is no longer
derived from the physical interface name (duplicated from ip_interface to be
specific) when ip_source=interface, and as a result, the interface must be set
explicitly.

Signed-off-by: Guoxin Pu <pugokushin@gmail.com>