luci-app-lxc: make regex for templates less strict
Starting with LXC v6.0.0, the way templates are parsed has changed [1].
As a result, trailing whitespace is added to each string except the last
one. However, if the last string (i.e., a timestamp) is shorter—such as
when using a different timestamp format—trailing whitespace will still
be added.

This commit updates the regular expression to handle such cases.

[1] https://github.com/lxc/lxc/commit/84cc3155af502db586f644f1092831155e45ccca

Signed-off-by: Richard Muzik <richard.muzik@turris.com>